requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
py-bcrypt==0.4
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
